requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-environ==0.10.0
environ==1.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
psycopg2==2.9.6
Pillow==9.5.0
